@@ -1,12 +1,12 @@
 name: _firebase_database_collection_vm
 description: For internal use of the Firestore and Realtime Database packages
-version: 0.0.2
+version: 0.0.2+1
 homepage: https://github.com/fluttercommunity/firebase_flutter_sdk
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.8.0 <3.0.0'
 
 dev_dependencies:
-  test: ^1.0.0
+  test: '>=1.0.0 <2.0.0'
   dart_quickcheck:
     git: https://github.com/long1eu/dart_quickcheck.git
