@@ -2,13 +2,13 @@ name: firebase_core_dart_example
 description: Example app for firebase_core_dart
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^0.1.3
-  firebase_core: ^0.4.4+2
+  firebase_core: ^0.5.0
   firebase_core_dart:
     path: ../
 
