@@ -1,19 +1,19 @@
 name: firebase_core_dart
 description: Dart implementation of firebase_core using firebase_core_vm
-version: 0.0.1
+version: 0.0.2
 homepage: https://github.com/fluttercommunity/firebase_dart_sdk/tree/develop/firebase_core/firebase_core_dart
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core_platform_interface: ^1.0.4
-  firebase_core_vm: ^0.0.4
-  connectivity: ^0.4.8+1
-  shared_preferences: ^0.5.6+2
-  rxdart: ^0.23.1
+  firebase_core_platform_interface: ^2.0.0
+  firebase_core_vm: ^0.0.5
+  connectivity: '>=0.4.9+2 <2.0.0'
+  shared_preferences: '>=0.5.6+2 <2.0.0'
+  rxdart: ^0.24.1
 
 dev_dependencies:
   flutter_test:
