@@ -2,14 +2,14 @@ name: firebase_core_vm_example
 description: Example app for firebase_core_vm
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^0.1.3
-  shared_preferences: ^0.5.6+2
-  connectivity: ^0.4.8+1
+  shared_preferences: '>=0.5.10 <2.0.0'
+  connectivity: '>=0.4.9 <2.0.0'
   firebase_core_vm:
     path: ../
 
