@@ -1,15 +1,14 @@
 name: firebase_core_vm
 description: Dart implementation of Firebase Core SDK that provides FirebaseApps used by all Firebase services
-version: 0.0.4
+version: 0.0.5
 homepage: https://github.com/fluttercommunity/firebase_flutter_sdk
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
 dependencies:
-  rxdart: ^0.23.1
-  meta: ^1.0.0
-  _firebase_internal_vm: ^0.0.2
+  rxdart: ^0.24.1
+  _firebase_internal_vm: ^0.0.2+1
 
 dev_dependencies:
-  test: ^1.0.0
+  test: '>=1.0.0 <2.0.0'
