@@ -5,37 +5,51 @@
 import 'dart:io';
 
 import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_core_dart/firebase_core_dart.dart';
 
 FirebaseOptions get firebaseOptions {
-  if (isDesktop) {
+  if (Platform.isLinux) {
+    return const FirebaseOptions(
+      apiKey: 'AIzaSyD9HeqeXUOXJh_DPDl211x8seUXlNmiJj0',
+      appId: '1:233259864964:linux:0034c73393cdd58c1d50ac24850d6d01f1e57aff',
+      messagingSenderId: '233259864964',
+      projectId: 'flutter-sdk',
+    );
+  } else if (Platform.isMacOS) {
     return const FirebaseOptions(
       apiKey: 'AIzaSyBQgB5s3n8WvyCOxhCws-RVf3C-6VnGg0A',
-      databaseURL: 'https://flutter-sdk.firebaseio.com',
-      projectID: 'flutter-sdk',
-      storageBucket: 'flutter-sdk.appspot.com',
-      gcmSenderID: '233259864964',
-      googleAppID:
-          '1:233259864964:macos:0bdc69800dd31cde15627229f39a6379865e8be1',
+      appId: '1:233259864964:macos:0bdc69800dd31cde15627229f39a6379865e8be1',
+      messagingSenderId: '233259864964',
+      projectId: 'flutter-sdk',
+    );
+  } else if (Platform.isWindows) {
+    return const FirebaseOptions(
+      apiKey: 'AIzaSyBNeYDWMlalWRL2M2_UhE5kiMmvVf3o9BM',
+      appId: '1:233259864964:windows:0034c73393cdd58c1d50ac24850d6d01f1e57aff',
+      messagingSenderId: '233259864964',
+      projectId: 'flutter-sdk',
+    );
+  } else if (Platform.isAndroid) {
+    return const FirebaseOptions(
+      apiKey: 'AIzaSyAM1bGAY-Bd4onFPFb2dBCJA3kx0eiWnSg',
+      appId: '1:233259864964:android:b2ec71b130a3170cd583d1',
+      messagingSenderId: '233259864964',
+      projectId: 'flutter-sdk',
+    );
+  } else if (Platform.isIOS) {
+    return const FirebaseOptions(
+      apiKey: 'AIzaSyBguTk4w2Xk2LD0mSdB2Pi9LTtt5BeAE6U',
+      appId: '1:233259864964:ios:fff621fea008bff1d583d1',
+      messagingSenderId: '233259864964',
+      projectId: 'flutter-sdk',
+    );
+  } else if (Platform.isFuchsia) {
+    return const FirebaseOptions(
+      apiKey: 'AIzaSyBOPFxmw3fni8Inzb_RhFDjb9zznXHfaRo',
+      appId: '1:233259864964:fuchsia:8fc440667cd119c335cf58c7cbfd4374f96fe786',
+      messagingSenderId: '233259864964',
+      projectId: 'flutter-sdk',
     );
   } else {
-    assert(isMobile);
-    return FirebaseOptions(
-      apiKey: Platform.isIOS
-          ? 'AIzaSyBguTk4w2Xk2LD0mSdB2Pi9LTtt5BeAE6U'
-          : 'AIzaSyAM1bGAY-Bd4onFPFb2dBCJA3kx0eiWnSg',
-      bundleID: 'eu.long1.firebaseCoreDartExample',
-      androidClientID:
-          '233259864964-atj096gj4dkn2q5iciufgrugequubseo.apps.googleusercontent.com',
-      clientID:
-          '233259864964-6agjde1utg6tml9bbe3uag4ppq8ogkb1.apps.googleusercontent.com',
-      googleAppID: Platform.isIOS
-          ? '1:233259864964:ios:875d94c61884160dd583d1'
-          : '1:233259864964:android:87954400984e5506d583d1',
-      databaseURL: 'https://flutter-sdk.firebaseio.com',
-      projectID: 'flutter-sdk',
-      storageBucket: 'flutter-sdk.appspot.com',
-      gcmSenderID: '233259864964',
-    );
+    throw UnimplementedError();
   }
 }
