@@ -1,7 +1,3 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_core_dart/firebase_core_dart.dart';
 import 'package:flutter/material.dart';
@@ -9,31 +5,59 @@ import 'package:flutter/material.dart';
 import 'firebase_config_js.dart' if (dart.library.io) 'firebase_config_io.dart';
 
 void main() {
-  if (isDesktop) {
-    FirebaseCoreDart.register();
+  if (kIsDesktop) {
+    FirebaseDart.register(options: firebaseOptions);
   }
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
-class MyApp extends StatelessWidget {
-  final String name = 'foo';
+class MyApp extends StatefulWidget {
+  const MyApp({Key key}) : super(key: key);
+
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  final String _name = 'foo';
+
+  String _message;
 
   Future<void> _configure() async {
-    final FirebaseApp app =
-        await FirebaseApp.configure(name: name, options: firebaseOptions);
-    assert(app != null);
-    print('Configured $app');
+    final bool exists = Firebase.apps.any((FirebaseApp app) => app.name == _name);
+    if (exists) {
+      final FirebaseApp app = Firebase.app(_name);
+
+      setState(() {
+        _message = 'App $_name already exists with ${app.options}.';
+      });
+    } else {
+      final FirebaseApp app = await Firebase.initializeApp(options: firebaseOptions, name: _name);
+
+      setState(() {
+        _message = 'App $_name was configured with ${app.options}';
+      });
+    }
   }
 
   Future<void> _allApps() async {
-    final List<FirebaseApp> apps = await FirebaseApp.allApps();
-    print('Currently configured apps: $apps');
+    setState(() {
+      _message = 'Currently configured apps: ${Firebase.apps}';
+    });
   }
 
   Future<void> _options() async {
-    final FirebaseApp app = await FirebaseApp.appNamed(name);
-    final FirebaseOptions options = await app?.options;
-    print('Current options for app $name: $options');
+    try {
+      final FirebaseApp app = Firebase.app(_name);
+      final FirebaseOptions options = app.options;
+      setState(() {
+        _message = 'Current options for app $_name: $options';
+      });
+    } catch (e) {
+      setState(() {
+        _message = 'The app has not been yet initialized.';
+      });
+    }
   }
 
   @override
@@ -49,10 +73,24 @@ class MyApp extends StatelessWidget {
             mainAxisAlignment: MainAxisAlignment.spaceAround,
             crossAxisAlignment: CrossAxisAlignment.stretch,
             children: <Widget>[
+              if (_message != null)
+                Expanded(
+                  child: Center(
+                    child: Text(_message),
+                  ),
+                ),
               RaisedButton(
-                  onPressed: _configure, child: const Text('initialize')),
-              RaisedButton(onPressed: _allApps, child: const Text('allApps')),
-              RaisedButton(onPressed: _options, child: const Text('options')),
+                onPressed: _configure,
+                child: const Text('initialize'),
+              ),
+              RaisedButton(
+                onPressed: _allApps,
+                child: const Text('allApps'),
+              ),
+              RaisedButton(
+                onPressed: _options,
+                child: const Text('options'),
+              ),
             ],
           ),
         ),
